requirements.txt
langchain
langchain-core
langgraph
langchain-google-genai
chromadb
psycopg2-binary
streamlit
pydantic